UI/requirements.txt
bidict==0.23.1
blinker==1.9.0
bottle==0.13.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-SocketIO==5.5.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
proxy_tools==0.1.0
pycparser==2.22
python-engineio==4.11.2
python-socketio==5.12.1
pythonnet==3.0.5
pywebview==5.4
requests==2.32.3
simple-websocket==1.1.0
typing_extensions==4.12.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wsproto==1.2.0
